typeファイルを切り出す
